Reworked some instructions, fixed GPU and added programmable interrupts

+++ VirtualPCCpp/Memory.cpp
@@ -6,14 +6,14 @@ Memory::Memory(int size) {
 }
 
 void Memory::Clear() {
-	for (int i = 0; i < memory.size(); i++) {
-		memory[i] = 0;
+	for (unsigned int i = 0; i < memory.size(); i++) {
+		memory[(int)i] = 0;
 	}
 }
 
 bool Memory::Empty() {
-	for (int i = 0; i < memory.size(); i++) {
-		if (memory[i] != 0) {
+	for (unsigned int i = 0; i < memory.size(); i++) {
+		if (memory[(int)i] != 0) {
 			return false;
 		}
 	}
